Agregar chat header
